+++ installation-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.1.0-RC1 (2016-05-XX)
 
+ * Pass the language and path to the Twig templates (see #15).
  * Fix the library loader and add the contao.cache service to the installation kernel (see #13).
 
 ### 1.1.0-beta1 (2016-04-25)